refact: remove self-implemented github service
