+++ zhang-cli/Cargo.toml
@@ -30,7 +30,7 @@ openssl = { workspace = true }
 env_logger = "0.11"
 self_update = "0.39"
 clap = { version = "4", features = ["derive"] }
-opendal = { version = "0.45", default-features = false, features = ["native-tls", "services-fs", "services-s3", "services-webdav"] }
+opendal = { version = "0.46", features = ["services-fs", "services-github", "services-webdav"] }
 
 async-recursion = "1.0.5"
 http2 = { package = "http", version = "0.2" }
