+++ zhang-core/src/process/close.rs
@@ -1,0 +1,31 @@
+use crate::ledger::Ledger;
+use crate::process::DirectiveProcess;
+use crate::{process, ZhangResult};
+use bigdecimal::Zero;
+use std::collections::HashMap;
+use zhang_ast::error::ErrorKind;
+use zhang_ast::{Close, SpanInfo};
+
+impl DirectiveProcess for Close {
+    fn validate(&mut self, ledger: &mut Ledger, span: &SpanInfo) -> ZhangResult<bool> {
+        let mut operations = ledger.operations();
+
+        // check if account exist
+        process::check_account_existed(self.account.name(), ledger, span)?;
+        process::check_account_closed(self.account.name(), ledger, span)?;
+
+        let balances = operations.single_account_latest_balances(self.account.name())?;
+        let has_non_zero_balance = balances.into_iter().any(|balance| !balance.balance_number.is_zero());
+        if has_non_zero_balance {
+            operations.new_error(ErrorKind::CloseNonZeroAccount, span, HashMap::default())?;
+        }
+        Ok(true)
+    }
+
+    fn process(&mut self, ledger: &mut Ledger, _span: &SpanInfo) -> ZhangResult<()> {
+        let mut operations = ledger.operations();
+
+        operations.close_account(self.account.name())?;
+        Ok(())
+    }
+}
